refactor: convert MFS APIs to async/await
Depends on:

* [ ] https://github.com/ipfs/js-ipfs-mfs/pull/66
