+++ src/core/components/stop.js
@@ -99,6 +99,7 @@ function createApi ({
   const api = {
     add,
     config: Commands.config({ repo }),
+    files: Commands.files({ ipld, blockService, repo, preload, options: constructorOptions }),
     init: () => { throw new AlreadyInitializedError() },
     start,
     stop: () => apiManager.api
